\name{BcellLymphomaCD79}
\alias{BcellLymphomaCD79}

\docType{data}
\title{Example data of bimodal CD79 expression.}

\description{Data set of CD79 expression in patients with B lymphoma (class 1) and in controls (class 2).}

\usage{data("BcellLymphomaCD79")}

\details{
Size 258429 x 2 , Dimensions 1, stored in \code{BcellLymphomaCD79$Data}

Classes 2, stored in \code{BcellLymphomaCD79$Classes}
}

\examples{
data(BcellLymphomaCD79)
str(BcellLymphomaCD79)
}
