\name{FeatureselectionData}
\alias{FeatureselectionData}

\docType{data}
\title{Example data with two groups for feature selection sexperiments with the Impact effet size measure.}

\description{Dataset with 2 classes and 20 variables that allow calss separation at varous degrees of difficulty.}

\usage{data("FeatureselectionData")}

\details{
Size 2000 x 20 , Dimensions 1, stored in \code{FeatureselectionData$Var0001,...,FeatureselectionData$Var0020}

Classes 2, stored in \code{FeatureselectionData$Classes}
}

\examples{
data(FeatureselectionData)
str(FeatureselectionData)
}
