\name{GMD}
\alias{GMD}
\title{Gini's mean difference}
\description{
calculates the value of Gini's mean difference.
}
\usage{
GMD(Data)
}
\arguments{
 \item{Data}{the data as a vector.}
}
\references{
 Lotsch, J., and Ultsch, A. (2019): 
 Impact -- An R Package for calculation and visualisation of the Impact distance and data distribution-shape based effect size measure.
}
\author{
  Jorn Lotsch and Alfred Ultsch
}
\keyword{Gini's mean difference}
