\name{ImpactEffectsize-package}
\alias{ImpactEffectsize-package}
\docType{package}
\title{
ImpactEffectsize-package
}
\description{
Calculation and visualization of the Impact effect size measure.
A non-parametric effect size measure capturing changes in central tendency or shape of data distributions for feature selection preceding machine-learning. The package provides the necessary functions to calculate and plot the Impact effect size measure between two groups.
}
\references{
 Lotsch, J., and Ultsch, A. (2019): 
 Impact -- An R Package for calculation and visualisation of the Impact distance and data distribution-shape based effect size measure.
}
\keyword{Impact}
\keyword{effect size}
\keyword{Gini's mean difference}

