\name{StocksFluctuation}
\alias{StocksFluctuation}

\docType{data}
\title{Example data of stock fluctuation.}

\description{Data set of Log ratios of daily changes of n =5,522 for 10 German stocks with low fluctuation (class 1) or high fluctuation (class 2).}

\usage{data("StocksFluctuation")}

\details{
Size 5522 x 2 , Dimensions 1, stored in \code{StocksFluctuation$logFluctuation}

Classes 2, stored in \code{StocksFluctuation$Classes}
}

\examples{
data(StocksFluctuation)
str(StocksFluctuation)
}
