\name{SameMeansData}
\alias{SameMeansData}

\docType{data}
\title{Example artificial data with two groups of same means but different data distribution shapes.}

\description{Dataset with 2 classes six variables were both classes have the same means but different shapes of the distribution.}

\usage{data("SameMeansData")}

\details{
Size 2000 x 7 , Dimensions 1, stored in \code{SameMeansData$NOchangeInMandS,...,SameMeansData$NegChi2andGauss}

Classes 2, stored in \code{SameMeansData$Classes}
}

\examples{
data(SameMeansData)
str(SameMeansData)
}
